Log supervisord to container stdout
The supervisor logs are not guaranteed to be captured by rsyslog if rsyslogd is
(for whatever reason) not able to start properly. This makes capturing the
supervisord logs from outside the supervisord process necessary. I think the
container stdout is a proper channel for this, as some container metrics should
be captured from the outside anyway.